ecommerce_backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
psycopg2-binary==2.9.9
Pillow==10.3.0
stripe==8.4.0
drf-spectacular==0.27.2
django-redis==5.4.0
celery==5.4.0
gunicorn==22.0.0
python-decouple==3.8
sentry-sdk>=1.35
